requirements.txt
streamlit>=1.28.0
python-docx>=0.8.11
requests>=2.31.0
python-dotenv>=1.0.0
groq>=0.4.0
pandas>=2.0.0
supabase==2.28.3
google-auth>=2.25.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
